+++ jetpack_vendor/automattic/jetpack-classic-theme-helper/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.1-alpha] - unreleased
+
+This is an alpha version! The changes listed here are not final.
+
+### Added
+- Custom Post Types: Added Restaurant Menu CPT files.
+
 ## [0.8.0] - 2024-12-16
 ### Added
 - Jetpack Testimonials: Ensuring functionality runs via the Classic Theme Helper package. [#40388]
@@ -177,6 +184,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Add wordpress folder on gitignore. [#37177]
 
+[0.8.1-alpha]: https://github.com/Automattic/jetpack-classic-theme-helper/compare/v0.8.0...v0.8.1-alpha
 [0.8.0]: https://github.com/Automattic/jetpack-classic-theme-helper/compare/v0.7.4...v0.8.0
 [0.7.4]: https://github.com/Automattic/jetpack-classic-theme-helper/compare/v0.7.3...v0.7.4
 [0.7.3]: https://github.com/Automattic/jetpack-classic-theme-helper/compare/v0.7.2...v0.7.3